rearranged the files and cleaned up the code
